+++ changelog.d/20240103_113250_rra_DM_42337.md
@@ -1,0 +1,3 @@
+### Bug fixes
+
+- Return prepuller configuration in the API response with snake-case fields rather than camel-case fields, as was originally intended.
