+++ controller/src/controller/services/source/base.py
@@ -11,8 +11,7 @@ from ...models.domain.docker import DockerReference
 from ...models.domain.image import MenuImage
 from ...models.domain.kubernetes import KubernetesNodeImage
 from ...models.domain.rspimage import RSPImage, RSPImageCollection
-from ...models.v1.prepuller import PrepulledImage
-from ...models.v1.prepuller_config import PrepullerConfig
+from ...models.v1.prepuller import PrepulledImage, PrepullerOptions
 
 __all__ = ["ImageSource"]
 
@@ -113,7 +112,7 @@ class ImageSource(metaclass=ABCMeta):
     @abstractmethod
     async def update_images(
         self,
-        prepull: PrepullerConfig,
+        prepull: PrepullerOptions,
         node_cache: Mapping[str, list[KubernetesNodeImage]],
     ) -> RSPImageCollection:
         """Update image information and determine what images to prepull.
