+++ docs/dev/api/controller.rst
@@ -69,9 +69,6 @@ This documentation therefore exists only to assist developers and code analysis 
 .. automodapi:: controller.models.v1.prepuller
    :include-all-objects:
 
-.. automodapi:: controller.models.v1.prepuller_config
-   :include-all-objects:
-
 .. automodapi:: controller.services.builder.fileserver
    :include-all-objects:
